Make Menu API accept TopLevelWindow
