+++ atom/browser/api/atom_api_browser_window.h
@@ -118,27 +118,4 @@ class BrowserWindow : public TopLevelWindow,
 
 }  // namespace atom
 
-namespace mate {
-
-template <>
-struct Converter<atom::NativeWindow*> {
-  static bool FromV8(v8::Isolate* isolate,
-                     v8::Local<v8::Value> val,
-                     atom::NativeWindow** out) {
-    // null would be tranfered to NULL.
-    if (val->IsNull()) {
-      *out = NULL;
-      return true;
-    }
-
-    atom::api::BrowserWindow* window;
-    if (!Converter<atom::api::BrowserWindow*>::FromV8(isolate, val, &window))
-      return false;
-    *out = window->window();
-    return true;
-  }
-};
-
-}  // namespace mate
-
 #endif  // ATOM_BROWSER_API_ATOM_API_BROWSER_WINDOW_H_
