+++ atom/browser/api/atom_api_top_level_window.cc
@@ -9,6 +9,7 @@
 
 #include "atom/browser/api/atom_api_browser_view.h"
 #include "atom/browser/api/atom_api_menu.h"
+#include "atom/browser/api/atom_api_web_contents.h"
 #include "atom/common/color_util.h"
 #include "atom/common/native_mate_converters/callback.h"
 #include "atom/common/native_mate_converters/file_path_converter.h"
