+++ lib/browser/api/top-level-window.js
@@ -17,4 +17,8 @@ TopLevelWindow.prototype._init = function () {
   }
 }
 
+TopLevelWindow.getFocusedWindow = () => {
+  return TopLevelWindow.getAllWindows().find((win) => win.isFocused())
+}
+
 module.exports = TopLevelWindow
